+++ uui-editor/src/plugins/linkPlugin/constants.ts
@@ -1,1 +1,4 @@
+import { ELEMENT_LINK } from '@udecode/plate-link';
+
+export const LINK_KEY = ELEMENT_LINK;
 export const LINK_TYPE = 'link';